First implementation of the niche model
Difference to NGs model:
Instead of determining a speciation time for each node when it's born
and putting it in a priority queue, a node to speciate is drawn
according to its speciation rate at every iteration.
This seems to yield a *different* process, because the proabilities of
generating an infinite tree are vastly lower than in NGs implementation.